chore: move attempt-related methods to `AttemptController`
